feat: split game logic into modules
